% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_impact}
\alias{calc_impact}
\title{Calculate impact}
\usage{
calc_impact(
  data,
  metric,
  impact_cols,
  by_cols = NA,
  method = "tobias",
  sort_by = "abs_impact",
  sort_order = -1
)
}
\arguments{
\item{data}{placeholder}

\item{metric}{placeholder}

\item{impact_cols}{placeholder}

\item{by_cols}{placeholder}

\item{method}{placeholder}

\item{sort_by}{placeholder}

\item{sort_order}{placeholder}
}
\value{
a tibble
}
\description{
Calculate impact
}
