% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_meta_impact}
\alias{calc_meta_impact}
\title{Calculates "meta impact" - the difference in impact for positive vs negative sentiment}
\usage{
calc_meta_impact(data, prefix = "excluded")
}
\arguments{
\item{data}{a data.table containing impact}
}
\value{
a tibble
}
\description{
Calculates "meta impact" - the difference in impact for positive vs negative sentiment
}
