% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drop_obs}
\alias{drop_obs}
\title{Helper function used to implement method=Tobias in impact calculation}
\usage{
drop_obs(data, drop_levels_dt, metric, by_cols)
}
\arguments{
\item{data}{a data.table}

\item{drop_levels_dt}{a data.table of levels to drop for the calculation}

\item{metric}{the target variable}

\item{by_cols}{by groups}
}
\value{
a data.table
}
\description{
Helper function used to implement method=Tobias in impact calculation
}
