% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixlist}
\alias{fixlist}
\title{process list elements in imported JSON}
\usage{
fixlist(x)

fixlist(x)
}
\arguments{
\item{x}{a list of elements that should be converted to lists}
}
\value{
list of lists

list of lists
}
\description{
process list elements in imported JSON

A helper function for processing list elements in imported JSON
}
