% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_sankey}
\alias{make_sankey}
\title{Make a sankey plot of new TA data}
\usage{
make_sankey(
  dat,
  filter_string = NULL,
  plot_title = NULL,
  save_plot = FALSE,
  plot_type = "sankey",
  filename = NULL,
  fig.scale = 1.2,
  fig.width = 8,
  fig.height = 7,
  sankey_vars = c("ta1_categories", "detail.categoryMain", "entity_type"),
  weight_by = NULL,
  valid_intent = c("complaint", "return", "praise", "other", "comparison", "suggestion",
    "recommendation"),
  remove_TA1_multi_cat = TRUE,
  n_threshold = 1000
)
}
\arguments{
\item{dat}{a data.table}

\item{filter_string}{string: filter statements for row selection. defaults to NULL for no filtering}

\item{plot_title}{string: title assigned to plot}

\item{save_plot}{logical: TRUE saves the file. Defaults to FALSE}

\item{plot_type}{character: currently only supports "sankey", which is the default}

\item{filename}{character: path and filename for the saved file. Only used if save_plot=TRUE. Defaults to NULL.}

\item{fig.scale}{scalar: figure scale. defaults to 1.2}

\item{fig.width}{scalar: figure width. defauts to 8}

\item{fig.height}{scalar: figure height. defaults to 6}

\item{sankey_vars}{string: the variables to be plotted. order is preserved. defaults to
c("ta1_categories", "detail.categoryMain", "entity_type"),}

\item{weight_by}{character: if TRUE node sizes are adjusted for values found in the 'weight' column.
defaults to FALSE}

\item{valid_intent}{a character vector of valid intents. used to remove hallucinations. defaults to
c("complaint", "return", "praise", "other", "comparison", "suggestion", "recommendation"),}

\item{remove_TA1_multi_cat}{logical: should multi-category TA1 labels be retained? defaults to TRUE}

\item{n_threshold}{scalar: sets the minimum sample size threshold for including an element in the Sankey plot. defaults to 1000}
}
\value{
a ggplot2 object
}
\description{
Make a sankey plot of new TA data
}
