% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_for_sankey}
\alias{prep_for_sankey}
\title{Create the flow dataset expected for sankey plot}
\usage{
prep_for_sankey(
  dat,
  filter_string = NULL,
  weight_nodesize = FALSE,
  sankey_vars = c("ta1_categories", "detail.categoryMain", "entity_type"),
  valid_intent = c("complaint", "return", "praise", "other", "comparison", "suggestion",
    "recommendation"),
  remove_TA1_multi_cat = TRUE,
  n_threshold = 1000
)
}
\arguments{
\item{dat}{a data.table}

\item{filter_string}{string: filter statements for row selection. defaults to NULL for no filtering}

\item{weight_nodesize}{logical: if TRUE node sizes are adjusted for values found in the 'weight' column.
defaults to FALSE}

\item{sankey_vars}{string: the variables to be plotted. order is preserved}

\item{valid_intent}{a character vector of valid intents. used to remove hallucinations.}

\item{remove_TA1_multi_cat}{logical: should multi-category TA1 labels be retained? defaults to TRUE}

\item{n_threshold}{scalar: sets the minimum sample size threshold for including an element in the Sankey plot. defaults to 1000}
}
\value{
a data.table suitable for Sankey plotting
}
\description{
Create the flow dataset expected for sankey plot
}
