% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_ta2_json}
\alias{read_ta2_json}
\title{Converting TA2 JSON to a data.table}
\usage{
read_ta2_json(path_to_file)
}
\arguments{
\item{path_to_file}{string: path to the JSON file}
}
\value{
a data.table
}
\description{
Converting TA2 JSON to a data.table
}
